Remove Time out
Remove the 60 seconds time out and replace it with the MAX_INTEGER Value. This will ensure make sure that after the user has authorized the payment, the app will wait until it gets back a response from vodacom.

After measuring the average time for time out without any interaction with the USSD notification is about 1 minute and 6 seconds.